python/requirements.txt
flask==2.3.2
flasgger==0.9.7.1
requests==2.32.5
pytz==2024.1
playwright==1.55
python-dotenv==1.0.0
langchain==1.0.2
langchain-community==0.4.1
langchain-core==1.0.1
langchain-chroma==1.0.0
chromadb==1.2.2
ollama==0.6.0
langchain-text-splitters==1.0.0
unstructured==0.18.15
jq==1.10.0
langchain-ollama==1.0.0